BET-4 MetadataPatterns added, uptm messages reworked; Phaser refactored

+++ CoreMessages/src/main/java/ru/splat/messages/proxyup/IdMessage.java
@@ -6,7 +6,7 @@ package ru.splat.messages.proxyup;
 public class IdMessage extends ProxyUPMessage {
     private final Long transactionId;
 
-    public IdMessage(Long transactionId, Long userId) {
+    public IdMessage(Long transactionId, Integer userId) {
         super(userId);
         this.transactionId = transactionId;
     }
