+++ CoreMessages/src/main/java/ru/splat/messages/proxyup/check/CheckRequest.java
@@ -6,7 +6,7 @@ import ru.splat.messages.proxyup.IdMessage;
  * ProxyUPMessage for UP asking for bet state.
  */
 public class CheckRequest extends IdMessage {
-    public CheckRequest(Long transactionId, Long userId) {
+    public CheckRequest(Long transactionId, Integer userId) {
         super(transactionId, userId);
     }
 }
